fixed money not being removed correctly after purchasing an upgrade

+++ Assets/CreditClicker/CreditClickerGame.cs
@@ -27,7 +27,7 @@ namespace Assets.CreditClicker
         public float baseIncomeSpeed;
         public float incomeSpeed;
 
-        public int interestAmount;
+        public float interestAmount;
 
 
         void Start()
